+++ packages/enhanced/src/lib/container/ModuleFederationPlugin.ts
@@ -11,6 +11,7 @@ import {
   composeKeyWithSeparator,
   type moduleFederationPlugin,
 } from '@module-federation/sdk';
+import { PrefetchPlugin } from '@module-federation/data-prefetch/cli';
 import { normalizeWebpackPath } from '@module-federation/sdk/normalize-webpack-path';
 import type { Compiler, WebpackPluginInstance } from 'webpack';
 import SharePlugin from '../sharing/SharePlugin';
@@ -20,6 +21,7 @@ import FederationRuntimePlugin from './runtime/FederationRuntimePlugin';
 import { RemoteEntryPlugin } from './runtime/RemoteEntryPlugin';
 import { ExternalsType } from 'webpack/declarations/WebpackOptions';
 import StartupChunkDependenciesPlugin from '../startup/MfStartupChunkDependenciesPlugin';
+import FederationModulesPlugin from './runtime/FederationModulesPlugin';
 
 const isValidExternalsType = require(
   normalizeWebpackPath(
@@ -64,7 +66,8 @@ class ModuleFederationPlugin implements WebpackPluginInstance {
         compiler,
       );
     }
-    if (options.experiments?.federationRuntime === 'hoisted') {
+    if (options.experiments?.federationRuntime) {
+      new FederationModulesPlugin().apply(compiler);
       new StartupChunkDependenciesPlugin({
         asyncChunkLoading: true,
       }).apply(compiler);
@@ -73,6 +76,7 @@ class ModuleFederationPlugin implements WebpackPluginInstance {
     if (options.dts !== false) {
       new DtsPlugin(options).apply(compiler);
     }
+    new PrefetchPlugin(options).apply(compiler);
 
     new FederationRuntimePlugin(options).apply(compiler);
 
@@ -124,6 +128,7 @@ class ModuleFederationPlugin implements WebpackPluginInstance {
           shareScope: options.shareScope,
           exposes: options.exposes!,
           runtimePlugins: options.runtimePlugins,
+          experiments: options.experiments,
         }).apply(compiler);
       }
       if (