+++ packages/enhanced/src/lib/container/runtime/FederationRuntimePlugin.ts
@@ -5,8 +5,9 @@ import type {
   Chunk,
 } from 'webpack';
 import { normalizeWebpackPath } from '@module-federation/sdk/normalize-webpack-path';
+import { PrefetchPlugin } from '@module-federation/data-prefetch/cli';
+import { moduleFederationPlugin } from '@module-federation/sdk';
 import FederationRuntimeModule from './FederationRuntimeModule';
-import type { moduleFederationPlugin } from '@module-federation/sdk';
 import {
   getFederationGlobalScope,
   normalizeRuntimeInitOptionsWithOutShared,
@@ -18,9 +19,15 @@ import fs from 'fs';
 import path from 'path';
 import { TEMP_DIR } from '../constant';
 import EmbedFederationRuntimePlugin from './EmbedFederationRuntimePlugin';
-import ContainerEntryModule from '../ContainerEntryModule';
+import FederationModulesPlugin from './FederationModulesPlugin';
 import HoistContainerReferences from '../HoistContainerReferencesPlugin';
 import pBtoa from 'btoa';
+import ContainerEntryDependency from '../ContainerEntryDependency';
+import FederationRuntimeDependency from './FederationRuntimeDependency';
+
+const ModuleDependency = require(
+  normalizeWebpackPath('webpack/lib/dependencies/ModuleDependency'),
+) as typeof import('webpack/lib/dependencies/ModuleDependency');
 
 const { RuntimeGlobals, Template } = require(
   normalizeWebpackPath('webpack'),
@@ -49,24 +56,29 @@ const EmbeddedRuntimePath = require.resolve(
 
 const federationGlobal = getFederationGlobalScope(RuntimeGlobals);
 
+const onceForCompler = new WeakSet();
+
 class FederationRuntimePlugin {
   options?: moduleFederationPlugin.ModuleFederationPluginOptions;
   entryFilePath: string;
   bundlerRuntimePath: string;
+  federationRuntimeDependency?: FederationRuntimeDependency; // Add this line
 
   constructor(options?: moduleFederationPlugin.ModuleFederationPluginOptions) {
     this.options = options ? { ...options } : undefined;
     this.entryFilePath = '';
     this.bundlerRuntimePath = BundlerRuntimePath;
+    this.federationRuntimeDependency = undefined; // Initialize as undefined
   }
 
   static getTemplate(
-    runtimePlugins: string[],
+    compiler: Compiler,
+    options: moduleFederationPlugin.ModuleFederationPluginOptions,
     bundlerRuntimePath?: string,
-    // keep so that the hash changes if experiemts change
     experiments?: moduleFederationPlugin.ModuleFederationPluginOptions['experiments'],
   ) {
     // internal runtime plugin
+    const runtimePlugins = options.runtimePlugins;
     const normalizedBundlerRuntimePath = normalizeToPosixPath(
       bundlerRuntimePath || BundlerRuntimePath,
     );
@@ -87,7 +99,6 @@ class FederationRuntimePlugin {
         runtimePluginNames.push(runtimePluginName);
       });
     }
-
     const embedRuntimeLines = Template.asString([
       `if(!${federationGlobal}.runtime){`,
       Template.indent([
@@ -111,7 +122,7 @@ class FederationRuntimePlugin {
       Template.indent([
         runtimePluginNames.length
           ? Template.asString([
-              `const pluginsToAdd = [`,
+              `var pluginsToAdd = [`,
               Template.indent(
                 runtimePluginNames.map(
                   (item) => `${item} ? (${item}.default || ${item})() : false,`,
@@ -132,27 +143,31 @@ class FederationRuntimePlugin {
         Template.indent([`${federationGlobal}.installInitialConsumes()`]),
         '}',
       ]),
+      PrefetchPlugin.addRuntime(compiler, {
+        name: options.name!,
+      }),
       '}',
     ]);
   }
 
   static getFilePath(
-    containerName: string,
-    runtimePlugins: string[],
+    compiler: Compiler,
+    options: moduleFederationPlugin.ModuleFederationPluginOptions,
     bundlerRuntimePath?: string,
     experiments?: moduleFederationPlugin.ModuleFederationPluginOptions['experiments'],
   ) {
+    const containerName = options.name;
     const hash = createHash(
       `${containerName} ${FederationRuntimePlugin.getTemplate(
-        runtimePlugins,
+        compiler,
+        options,
         bundlerRuntimePath,
         experiments,
       )}`,
     );
     return path.join(TEMP_DIR, `entry.${hash}.js`);
   }
-
-  getFilePath() {
+  getFilePath(compiler: Compiler) {
     if (this.entryFilePath) {
       return this.entryFilePath;
     }
@@ -163,15 +178,16 @@ class FederationRuntimePlugin {
 
     if (!this.options?.virtualRuntimeEntry) {
       this.entryFilePath = FederationRuntimePlugin.getFilePath(
-        this.options.name!,
-        this.options.runtimePlugins!,
+        compiler,
+        this.options,
         this.bundlerRuntimePath,
         this.options.experiments,
       );
     } else {
       this.entryFilePath = `data:text/javascript;charset=utf-8;base64,${pBtoa(
         FederationRuntimePlugin.getTemplate(
-          this.options.runtimePlugins!,
+          compiler,
+          this.options,
           this.bundlerRuntimePath,
           this.options.experiments,
         ),
@@ -179,12 +195,11 @@ class FederationRuntimePlugin {
     }
     return this.entryFilePath;
   }
-
-  ensureFile() {
+  ensureFile(compiler: Compiler) {
     if (!this.options) {
       return;
     }
-    const filePath = this.getFilePath();
+    const filePath = this.getFilePath(compiler);
     try {
       fs.readFileSync(filePath);
     } catch (err) {
@@ -192,7 +207,8 @@ class FederationRuntimePlugin {
       fs.writeFileSync(
         filePath,
         FederationRuntimePlugin.getTemplate(
-          this.options.runtimePlugins!,
+          compiler,
+          this.options,
           this.bundlerRuntimePath,
           this.options.experiments,
         ),
@@ -200,27 +216,76 @@ class FederationRuntimePlugin {
     }
   }
 
+  getDependency(compiler: Compiler) {
+    if (this.federationRuntimeDependency)
+      return this.federationRuntimeDependency;
+    this.federationRuntimeDependency = new FederationRuntimeDependency(
+      this.getFilePath(compiler),
+    );
+    return this.federationRuntimeDependency;
+  }
+
   prependEntry(compiler: Compiler) {
     if (!this.options?.virtualRuntimeEntry) {
-      this.ensureFile();
+      this.ensureFile(compiler);
     }
-    const entryFilePath = this.getFilePath();
 
-    modifyEntry({
-      compiler,
-      prependEntry: (entry) => {
-        Object.keys(entry).forEach((entryName) => {
-          const entryItem = entry[entryName];
-          if (!entryItem.import) {
-            // TODO: maybe set this variable as constant is better https://github.com/webpack/webpack/blob/main/lib/config/defaults.js#L176
-            entryItem.import = ['./src'];
-          }
-          if (!entryItem.import.includes(entryFilePath)) {
-            entryItem.import.unshift(entryFilePath);
-          }
-        });
-      },
-    });
+    //if using runtime experiment, use the new include method else patch entry
+    if (this.options?.experiments?.federationRuntime) {
+      compiler.hooks.thisCompilation.tap(
+        this.constructor.name,
+        (compilation: Compilation, { normalModuleFactory }) => {
+          compilation.dependencyFactories.set(
+            FederationRuntimeDependency,
+            normalModuleFactory,
+          );
+          compilation.dependencyTemplates.set(
+            FederationRuntimeDependency,
+            new ModuleDependency.Template(),
+          );
+        },
+      );
+      compiler.hooks.make.tapAsync(
+        this.constructor.name,
+        (compilation: Compilation, callback) => {
+          const federationRuntimeDependency = this.getDependency(compiler);
+          const logger = compilation.getLogger('FederationRuntimePlugin');
+          const hooks =
+            FederationModulesPlugin.getCompilationHooks(compilation);
+          compilation.addInclude(
+            compiler.context,
+            federationRuntimeDependency,
+            { name: undefined },
+            (err, module) => {
+              if (err) {
+                return callback(err);
+              }
+              hooks.addFederationRuntimeModule.call(
+                federationRuntimeDependency,
+              );
+              callback();
+            },
+          );
+        },
+      );
+    } else {
+      const entryFilePath = this.getFilePath(compiler);
+      modifyEntry({
+        compiler,
+        prependEntry: (entry) => {
+          Object.keys(entry).forEach((entryName) => {
+            const entryItem = entry[entryName];
+            if (!entryItem.import) {
+              // TODO: maybe set this variable as constant is better https://github.com/webpack/webpack/blob/main/lib/config/defaults.js#L176
+              entryItem.import = ['./src'];
+            }
+            if (!entryItem.import.includes(entryFilePath)) {
+              entryItem.import.unshift(entryFilePath);
+            }
+          });
+        },
+      });
+    }
   }
 
   injectRuntime(compiler: Compiler) {
@@ -238,11 +303,6 @@ class FederationRuntimePlugin {
     compiler.hooks.thisCompilation.tap(
       this.constructor.name,
       (compilation: Compilation, { normalModuleFactory }) => {
-        const isEnabledForChunk = (chunk: Chunk): boolean => {
-          const [entryModule] =
-            compilation.chunkGraph.getChunkEntryModulesIterable(chunk) || [];
-          return entryModule instanceof ContainerEntryModule;
-        };
         const handler = (chunk: Chunk, runtimeRequirements: Set<string>) => {
           if (runtimeRequirements.has(federationGlobal)) return;
           runtimeRequirements.add(federationGlobal);
@@ -370,18 +430,21 @@ class FederationRuntimePlugin {
         },
       );
     }
+
     if (this.options?.experiments?.federationRuntime === 'hoisted') {
       this.bundlerRuntimePath = this.bundlerRuntimePath.replace(
         '.cjs.js',
         '.esm.js',
       );
+
       new EmbedFederationRuntimePlugin(this.bundlerRuntimePath).apply(compiler);
 
       new HoistContainerReferences(
         this.options.name ? this.options.name + '_partial' : undefined,
         // hoist all modules of federation entry
-        this.getFilePath(),
+        this.getFilePath(compiler),
         this.bundlerRuntimePath,
+        this.options.experiments,
       ).apply(compiler);
 
       new compiler.webpack.NormalModuleReplacementPlugin(
@@ -397,9 +460,13 @@ class FederationRuntimePlugin {
         },
       ).apply(compiler);
     }
-    this.prependEntry(compiler);
-    this.injectRuntime(compiler);
-    this.setRuntimeAlias(compiler);
+    // dont run multiple times on every apply()
+    if (!onceForCompler.has(compiler)) {
+      this.prependEntry(compiler);
+      this.injectRuntime(compiler);
+      this.setRuntimeAlias(compiler);
+      onceForCompler.add(compiler);
+    }
   }
 }