+++ packages/bridge/vue3-bridge/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @module-federation/bridge-vue3
 
+## 0.6.6
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.6.6
+
+## 0.6.5
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.6.5
+
+## 0.6.4
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.6.4
+
+## 0.6.3
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.6.3
+
+## 0.6.2
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.6.2
+
 ## 0.6.1
 
 ### Patch Changes