+++ packages/modernjs/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @module-federation/modern-js
 
+## 0.6.6
+
+### Patch Changes
+
+- @module-federation/enhanced@0.6.6
+- @module-federation/node@2.5.16
+- @module-federation/sdk@0.6.6
+
+## 0.6.5
+
+### Patch Changes
+
+- @module-federation/enhanced@0.6.5
+- @module-federation/node@2.5.15
+- @module-federation/sdk@0.6.5
+
+## 0.6.4
+
+### Patch Changes
+
+- @module-federation/enhanced@0.6.4
+- @module-federation/node@2.5.14
+- @module-federation/sdk@0.6.4
+
+## 0.6.3
+
+### Patch Changes
+
+- 81201b8: fix(modernjs): mfConfigPlugin should run after @modern-js/plugin-initialize
+  - @module-federation/enhanced@0.6.3
+  - @module-federation/sdk@0.6.3
+  - @module-federation/node@2.5.13
+
+## 0.6.2
+
+### Patch Changes
+
+- 541494d: fix(modernjs): correct splitChunks.cacheGroups key which need to be removed
+- 2394e38: fix(modernjs): auto set enableAsyncEntry when bundler is rspack
+  - @module-federation/node@2.5.12
+  - @module-federation/enhanced@0.6.2
+  - @module-federation/sdk@0.6.2
+
 ## 0.6.1
 
 ### Patch Changes