+++ packages/rspack/CHANGELOG.md
@@ -1,5 +1,62 @@
 # @module-federation/rspack
 
+## 0.6.6
+
+### Patch Changes
+
+- Updated dependencies [35aead4]
+  - @module-federation/dts-plugin@0.6.6
+  - @module-federation/manifest@0.6.6
+  - @module-federation/sdk@0.6.6
+  - @module-federation/runtime-tools@0.6.6
+  - @module-federation/managers@0.6.6
+  - @module-federation/bridge-react-webpack-plugin@0.6.6
+
+## 0.6.5
+
+### Patch Changes
+
+- @module-federation/sdk@0.6.5
+- @module-federation/runtime-tools@0.6.5
+- @module-federation/managers@0.6.5
+- @module-federation/manifest@0.6.5
+- @module-federation/dts-plugin@0.6.5
+- @module-federation/bridge-react-webpack-plugin@0.6.5
+
+## 0.6.4
+
+### Patch Changes
+
+- Updated dependencies [d90295b]
+  - @module-federation/bridge-react-webpack-plugin@0.6.4
+  - @module-federation/sdk@0.6.4
+  - @module-federation/runtime-tools@0.6.4
+  - @module-federation/managers@0.6.4
+  - @module-federation/manifest@0.6.4
+  - @module-federation/dts-plugin@0.6.4
+
+## 0.6.3
+
+### Patch Changes
+
+- @module-federation/sdk@0.6.3
+- @module-federation/runtime-tools@0.6.3
+- @module-federation/managers@0.6.3
+- @module-federation/manifest@0.6.3
+- @module-federation/dts-plugin@0.6.3
+- @module-federation/bridge-react-webpack-plugin@0.6.3
+
+## 0.6.2
+
+### Patch Changes
+
+- @module-federation/dts-plugin@0.6.2
+- @module-federation/runtime-tools@0.6.2
+- @module-federation/sdk@0.6.2
+- @module-federation/managers@0.6.2
+- @module-federation/manifest@0.6.2
+- @module-federation/bridge-react-webpack-plugin@0.6.2
+
 ## 0.6.1
 
 ### Patch Changes