+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @module-federation/runtime
 
+## 0.6.6
+
+### Patch Changes
+
+- @module-federation/sdk@0.6.6
+
+## 0.6.5
+
+### Patch Changes
+
+- @module-federation/sdk@0.6.5
+
+## 0.6.4
+
+### Patch Changes
+
+- @module-federation/sdk@0.6.4
+
+## 0.6.3
+
+### Patch Changes
+
+- @module-federation/sdk@0.6.3
+
+## 0.6.2
+
+### Patch Changes
+
+- 9f98292: fix(runtime): set loading if registeredShared not set
+  - @module-federation/sdk@0.6.2
+
 ## 0.6.1
 
 ### Patch Changes